fix: Windows compat, namespace collision, and multiplatform CI (#50)
- Replace Unicode arrow (→) with ASCII (->) in benchmark prints to fix
  UnicodeEncodeError on Windows cp1252 consoles
- Move benchmark entry point from top-level main.py to sakura/benchmark.py
  to avoid polluting site-packages global namespace
- Update sakura-benchmark script to point to sakura.benchmark:main
- Relax charset-normalizer pin from ==3.1.0 to >=3.1.0
- Extend CI test matrix to run on ubuntu, macos, and windows

Closes #50
